Clase Ataque hecha abstracta. AtaqueBasico y AtaqueEspecial heredan de Ataque
