+++ src/Library/Contenido Parte II/LeerArchivo.cs
@@ -1,4 +1,5 @@
 using System.Globalization;
+using Library.Contenido_Parte_II;
 
 namespace Proyecto_Pokemones_I;
 
@@ -96,7 +97,7 @@ public static class LeerArchivo
                 AtaqueEspecial ataque4 = new AtaqueEspecial(ataqueNombre, ataqueTipo, ataqueDaño, ataquePrecision, ataqueEfecto);
 
                 //Define el atributo Lista de Ataques del Pokemon:
-                List<IAtaque> pokeAtaques = new List<IAtaque>();
+                List<Ataque> pokeAtaques = new List<Ataque>();
                 pokeAtaques.Add(ataque1);
                 pokeAtaques.Add(ataque2);
                 pokeAtaques.Add(ataque3);
