+++ src/Library/Contenido Parte II/Pokemon.cs
@@ -9,7 +9,7 @@ public class Pokemon
     private double vida;
     private double vidaMax;
     private double velocidadAtaque;
-    private List<IAtaque> listadoAtaques;
+    private List<Ataque> listadoAtaques;
     public string EfectoActivo { get; set;}
     public bool PuedeAtacar { get; set; }
     public int TurnosDuracionEfecto { get; set; }
@@ -35,7 +35,7 @@ public class Pokemon
     {
         return this.velocidadAtaque;
     }
-    public List<IAtaque> GetAtaques()
+    public List<Ataque> GetAtaques()
     {
         return this.listadoAtaques;
     }
@@ -43,7 +43,7 @@ public class Pokemon
     
 
     //Constructor:
-    public Pokemon(string pokeNombre, string pokeTipo, double pokeVida, double pokeVelAtaque, List<IAtaque> ataques)
+    public Pokemon(string pokeNombre, string pokeTipo, double pokeVida, double pokeVelAtaque, List<Ataque> ataques)
     {
         this.nombre = pokeNombre;
         this.tipo = pokeTipo;
@@ -62,7 +62,7 @@ public class Pokemon
         this.vida = vida-dañoEspecial;
     }
     
-    public void RecibirDaño(IAtaque ataqueRecibido)
+    public void RecibirDaño(Ataque ataqueRecibido)
     {
         DiccionarioTipos.GetInstancia();
         List<string> listaDebilidades = DiccionarioTipos.GetDebilContra(this.tipo);
