+++ test/LibraryTests/TestUserStory7.cs
@@ -1,3 +1,4 @@
+using Library.Contenido_Parte_II;
 using NUnit.Framework;
 using Proyecto_Pokemones_I;
 
@@ -32,7 +33,7 @@ public class TestUserStory7
         
         AtaqueBasico rayo = new AtaqueBasico("RAYO", "ELÉCTRICO", 10000, 100);
         AtaqueBasico hidrobomba = new AtaqueBasico("HIDROBOMBA", "AGUA", 10, 100);
-        List<IAtaque> pikachuataques = new List<IAtaque>();
+        List<Ataque> pikachuataques = new List<Ataque>();
         
         pikachuataques.Add(rayo);
         pikachuataques.Add(hidrobomba);
